refactor: Start moving cache management code into one place. This commit is for the AtRefreshJob * Created CacheManager class, to be responsible for getting from cache and updating cache * Inject a CacheManager instance into AtRefreshJob via its constructor * Migrate all cache retrieves and updates from AtRefreshJob into the CacheManager * AtSecondaryServerImpl constructs CacheManager instance and then injects into the AtRefreshJob instance
