+++ packages/at_secondary_server/lib/src/caching/at_refresh_job.dart
@@ -1,102 +1,30 @@
-import 'package:at_commons/at_commons.dart';
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
-import 'package:at_secondary/src/connection/inbound/dummy_inbound_connection.dart';
-import 'package:at_secondary/src/connection/outbound/outbound_client_manager.dart';
+import 'package:at_secondary/src/caching/cache_manager.dart';
 import 'package:at_utils/at_logger.dart';
 import 'package:cron/cron.dart';
 
 class AtRefreshJob {
   final String atSign;
-  final SecondaryKeyStore<String, AtData?, AtMetaData?> keyStore;
   late Cron _cron;
-  final OutboundClientManager outboundClientManager;
+  final CacheManager cacheManager;
 
-  AtRefreshJob(this.atSign, this.keyStore, this.outboundClientManager);
+  AtRefreshJob(this.atSign, this.cacheManager);
 
   final logger = AtSignLogger('AtRefreshJob');
 
-  /// Returns the list of cached keys
-  Future<List<String>> _getCachedKeys() async {
-    List<String> keysList = keyStore.getKeys(regex: CACHED);
-    var cachedKeys = <String>[];
-    var now = DateTime.now().toUtc();
-    var nowInEpoch = now.millisecondsSinceEpoch;
-    var itr = keysList.iterator;
-    while (itr.moveNext()) {
-      var key = itr.current;
-      AtMetaData? metadata = await keyStore.getMeta(key);
-      // Setting metadata.ttr = -1 represents not to updated the cached key.
-      // Hence skipping the key from refresh job.
-      if (metadata == null || metadata.ttr == -1) {
-        continue;
-      }
-      if (metadata.ttr != -1 &&
-          metadata.refreshAt!.millisecondsSinceEpoch > nowInEpoch) {
-        continue;
-      }
-      // If metadata.availableAt is greater is lastRefreshedAtInEpoch, key's TTB is not met.
-      if (metadata.availableAt != null &&
-          metadata.availableAt!.millisecondsSinceEpoch >= nowInEpoch) {
-        continue;
-      }
-      // If metadata.expiresAt is less than nowInEpoch, key's TTL is expired.
-      if (metadata.expiresAt != null &&
-          nowInEpoch >= metadata.expiresAt!.millisecondsSinceEpoch) {
-        continue;
-      }
-      cachedKeys.add(key);
-    }
-    return cachedKeys;
-  }
-
-  /// Returns of the value of the key from the another secondary server.
-  /// Key to lookup on the another secondary server.
-  /// Future<String> value of the key.
-  Future<String?> _lookupValue(String key, {bool isHandShake = true}) async {
-    var index = key.indexOf('@');
-    var otherAtSign = key.substring(index);
-    var outBoundClient = outboundClientManager.getClient(
-        otherAtSign, DummyInboundConnection(),
-        isHandShake: isHandShake)!;
-    // Need not connect again if the client's handshake is already done
-
-    if (!outBoundClient.isHandShakeDone) {
-      var connectResult =
-      await outBoundClient.connect(handshake: isHandShake);
-      logger.finer('connect result: $connectResult');
-    }
-    return await outBoundClient.lookUp(key, handshake: isHandShake);
-  }
-
-  /// Updates the cached key with the new value.
-  Future<void> _updateCachedValue(
-      String? newValue, AtData? oldValue, var cachedKeyName) async {
-    var atData = AtData();
-    atData.data = newValue;
-    atData.metaData = oldValue?.metaData;
-    await keyStore.put(cachedKeyName, atData);
-  }
-
   /// The refresh job
   Future<void> _refreshJob(int runFrequencyHours) async {
-    var keysToRefresh = await _getCachedKeys();
-    String lookupKey;
+    var keysToRefresh = await cacheManager.getCachedKeys();
+
     var itr = keysToRefresh.iterator;
     while (itr.moveNext()) {
       var cachedKeyName = itr.current;
-      lookupKey = cachedKeyName;
       String? newValue;
 
       try {
-        if (lookupKey.startsWith('cached:public:')) {
-          lookupKey = lookupKey.replaceAll('cached:public:', '');
-          newValue = await _lookupValue(lookupKey, isHandShake: false);
-        } else {
-          lookupKey = lookupKey.replaceAll('$CACHED:$atSign:', '');
-          newValue = await _lookupValue(lookupKey);
-        }
+          newValue = await cacheManager.lookUpRemoteValue(cachedKeyName);
       } catch (e) {
-        logger.info("Exception while looking up $lookupKey : $e");
+        logger.info("Exception while trying to get latest value for $cachedKeyName : $e");
         continue;
       }
       // If new value is null, do nothing. Continue for next key.
@@ -108,22 +36,20 @@ class AtRefreshJob {
       // do not update the cached key. Do nothing. Continue for next key.
       if (newValue.trim().isEmpty || newValue == 'null') {
         logger.finest(
-            'value not found for $lookupKey. Failed updating the cached key');
+            'value not found for $cachedKeyName. Failed updating the cached key');
         continue;
       }
       // If old value and new value are equal, then do not update;
       // Continue for next key.
-      var oldValue = await keyStore.get(cachedKeyName);
+      AtData? oldValue = await cacheManager.getCachedValue(cachedKeyName);
       if (oldValue?.data == newValue) {
         logger.finest(
-            '$lookupKey cached value is same as looked-up value. Not updating the cached key');
+            '$cachedKeyName cached value is same as looked-up value. Not updating the cached key');
         continue;
       }
-      logger.finest('Updated the cached key value of $lookupKey with $newValue');
-      await _updateCachedValue(newValue, oldValue, cachedKeyName);
-      //Update the refreshAt date for the next interval.
-      var atMetadata = AtMetadataBuilder(ttr: oldValue!.metaData!.ttr).build();
-      await keyStore.putMeta(cachedKeyName, atMetadata);
+
+      await cacheManager.updateCachedValue(newValue, oldValue, cachedKeyName);
+      logger.finer('Updated the cached key value of $cachedKeyName with $newValue');
     }
   }
